Выделены save_depot_config и save_part_depot_config.
Методы временно выделены в отдельные файлы для упрощения обработки.
